Better memory handling (doubles is now floats.. We don't use more than 2 decimal precision and never around 2 billion) We now use a Node-array instead of a Map (for implementing a List essentially), we can do this because we know how many nodes to load. Loading time is better now.

+++ Map/src/dk/itu/groupe/Node.java
@@ -11,8 +11,8 @@ import dk.itu.groupe.loading.DataLine;
 public class Node
 {
     private final int id;
-    private final double x;
-    private final double y;
+    private final float x;
+    private final float y;
 
     /**
      * Parses node data from line, throws an IOException if something unexpected
@@ -24,8 +24,8 @@ public class Node
     {
         DataLine dl = new DataLine(line);
         id = dl.getInt();
-        x = dl.getDouble();
-        y = dl.getDouble();
+        x = dl.getFloat();
+        y = dl.getFloat();
     }
     
     public int id()
@@ -33,12 +33,12 @@ public class Node
         return id;
     }
     
-    public double x()
+    public float x()
     {
         return x;
     }
     
-    public double y()
+    public float y()
     {
         return y;
     }
